[CLEANUP] Move getApplicationRoot to a separate class
This method needs to be (indirectly) called from the ApplicationKernel
constructor, making setting the application root from the Bootstrap class
impossible.

In addition, this method will be needed in the Composer ScriptHandler class,
which should not need to bootstrap the application.
